pyproject.toml
[tool.poetry]
name = "api-devsolution-db"
version = "0.1.0"
description = ""
authors = ["keitarods <83796605+keitarods@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.28"
fastapi = "^0.110.0"
httpx = "^0.27.0"
pydantic = "^2.6.3"
uvicorn = "^0.28.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
